build: Fix optional binman symbols being required with LTO
Binman has a feature to fill in symbols declared on the C side with
values about entries in the images it builds. Binman raises an error if
it cannot find an entry referred by these symbols, unless the symbol is
declared as optional by marking it as 'weak'.

For reasons I do not understand, enabling CONFIG_LTO removes this 'weak'
attribute from optional binman symbols, causing binman to raise an error
when it should only print a warning.

This patch re-declares the symbols in a generated keep-syms-lto.c file,
which somehow makes the symbols 'weak' again. One issue here is that I
don't know how to get the original type of the symbol, so all of them
are re-declared as 'ulong'.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>