rk3399: kevin: Do not reset on panic
Normally U-Boot immediately resets the board on a panic. It has been
possible to run into "Synchronous Abort"s from shim and rEFInd due to
other issues, and having a traceback in those cases can be useful. Hang
instead of resetting for chromebook_kevin, so the panic reason stays on
the screen.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>