video: test: Test partial updates of hardware frame buffer
With VIDEO_COPY enabled, only the modified parts of the frame buffer are
intended to be copied to the hardware. Add a test that checks this, by
overwriting contents we prepared without telling the video uclass and
then checking if the overwritten contents have been redrawn on the next
sync.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Series-changes: 5
- Add patch "video: test: Test partial updates of hardware frame buffer"

Commit-changes: 6
- Add tag "Reviewed-by: Simon Glass <sjg@chromium.org>"