video: test: Test video damage tracking via vidconsole
With VIDEO_DAMAGE, the video uclass tracks updated regions of the frame
buffer in order to avoid unnecessary work during a video sync. Enable
the config in sandbox and add a test for it, by printing strings at a
few locations and checking the tracked region.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>
Commit-notes:
This is hard to test because most things issue video syncs that process
and reset the damaged region.
END
Series-changes: 5
- Add patch "video: test: Test video damage tracking via vidconsole"