rk3399: kevin: Disable dcache in pre-relocation board early init
On chromebook_kevin, U-Boot runs very slow when started from depthcharge
as a RW_LEGACY payload. For unclear reasons, disabling then re-enabling
dcache via the prompt restores it back to a tolerable speed, although it
reports dcache status as enabled while it was slow.

Disable dcache in pre-relocation board-specific init hook if it happens
to be already enabled. It is later re-enabled post-relocation in the
common init sequence.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>