RFC: sandbox: video: Use partial updates for SDL display
Now that we have video damage tracking, try to reduce the SDL display
work by copying only the updated regions onto the SDL texture instead of
the entire framebuffer. We still have to do RenderClear and RenderCopy
the whole texture onto the renderer, but that allegedly happens in the
GPU.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>
Commit-notes:
The second half of copy_to_texture is untested.
END
Series-name: sandbox-sdl-partial-render
Series-to: u-boot@lists.denx.de
Series-cc: Simon Glass <sjg@chromium.org>
Series-cc: Alper Nebi Yasak <alpernebiyasak@gmail.com>
Series-cc: Anatolij Gustschin <agust@denx.de>
Series-cc: Heinrich Schuchardt <xypron.glpk@gmx.de>
Cover-letter:
sandbox: video: Refactor out of uclass, try partial screen updates

While working on the video damage tracking series [1], I noticed the
10Hz limitation on the sandbox screen refresh rate, and thought maybe
applying the damage tracking idea to SDL would make it reasonable to
increase the rate or even remove the limit completely.

Experimenting on that idea resulted in a bit of a refactoring to pull
sandbox out of the video uclass, so I'm sending it as a series in case
some of it could be useful.

As for the partial update idea itself, I found that having the delay
around 1/refresh-rate appears to be the most visually smooth (about
7ms for 144Hz, 16ms for 60Hz etc.), but couldn't exactly evaluate CPU
usage for which the original 10Hz rate limit was set. In any case, the
visual appeal of sandbox doesn't really matter, so I don't exactly care.

Of course, this applies onto the video damage tracking series that I
recently updated and sent [1], but only the "partial updates" patch
is functionally dependent on it.

[1] https://lore.kernel.org/u-boot/20230821135111.3558478-1-alpernebiyasak@gmail.com/
END