x86: qemu: Move qfw_pio device declaration to device-tree
The qfw_pio device for x86 QEMU virtual machines is statically declared
in arch-specific files with U_BOOT_DRVINFO(). Instead of this, add a
"qemu,fw-cfg-pio" compatible match to the qfw_pio driver and add a
corresponding node to the in-tree device-trees we have for x86 QEMU
boards.

QEMU firmware config drivers need to be probed to bind the ramfb device.
The ramfb driver needs to be bound before relocation to properly reserve
video memory for it, otherwise it cannot be probed after relocation. Add
the bootph-some-ram property to the new node so that ramfb can work as
an initial vidconsole.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>
Series-changes: 3
- Add patch "x86: qemu: Move QFW_PIO device declaration to device-tree"