RFC: video: Support using a cyclic function to call video sync
Functions that draw to the framebuffer usually call video_sync() to
ensure what they have drawn is actually visible on the display hardware.
Doing it for every little change (like for each character on the screen)
can end up being slow, but not doing so risks content never ending up on
the screen.

Implement and register a cyclic function that triggers a video sync, at
a configurable rate defaulting to 60Hz. This ensures each change will
eventually end up on the screen, and hypothetically provides an ideal
batching strategy for framebuffer changes.

Enabling this causes transient graphical artifacts on chromebook_kevin.
As far as I can tell, it's because the cache doesn't wait for us to
explicitly flush the changes to the hardware framebuffer. Implementing
and enabling copy framebuffer functionality for the driver solves most
of the artifacts, but also slows things down a bit.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>