HACK: rk3399: gru: Remove assigned clock dt properties for EDP node
Having the PCLK_EDP clock in the assigned-clocks property of the `edp`
node means that U-Boot tries to set its rate automatically. This clock
isn't implemented for the RK3399 clock driver, so it fails and prevents
display from being initialized.

The display happens to work fine without it, remove the property until
the clock driver can handle the clock.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>